outlier setup
IQR is the interquartile range or the distance between the first and the third quartile.

This way, we can define outliers in several ways:

a)Any value smaller than -1.5*IQR and greater than +1.5*IQR is an outlier. Sometimes, the coefficient is set to 2 or 3. All values between 1.5*IQR and 3*IQR are called mean outliers and the values above 3*IQR are called extreme outliers.

b)Any value that appears to be outside of the 5th and 95th percentile can be considered to be an outlier,

c)Points plotted three or more MSD away are also outliers.