Imputing NA instead of outliers
Imputation is a substitution of missing, incorrect or invalid values with other values. The input data for training the model must contain only valid values. You can either:

substitute NA for the mean, median, mod (the statistical characteristics of the set will not change),
substitute the outliers greater than 1.5*IQR for 0.95 percentile and the outliers smaller than - 1.5*IQR for 0.05 percentile.