Emitted emissions and normalized data
Previously, we determined that a set of data with emitted emissions and normalized data makes it possible to obtain better results in training with pre-training. We will use it. You can check other preprocessing options, which we considered earlier.

When dividing into pretrain / train / val / test-sections, we use the first opportunity to improve the quality of classification - we will increase the number of examples for training. In the pretrain section, the number of examples will be increased to 4000