Ranked Predictor Poset
We test the function clusterSim :: HINoV.Mod () (for a detailed description, see the package). It ranks the variables based on clustering with different distances and methods. We will use the default parameters. You can experiment with other parameters. The constant numFeature <- 10 makes it possible to change the number of the best predictors bestF supplied by the model.