Creation, training and testing of ensemble
We train the ensemble of neural networks (n <- 500 pieces), combining them in Ens. Each neural network is trained on its own sample. The sample is formed by extracting from the training set 7/10 examples randomly with a replacement. For the model, you need to set two parameters: nh - the number of neurons in the hidden layer and act - the activation function. The package offers the following options for activation functions:

- sig: sigmoid
- sin: sine
- radbas: radial basis
- hardlim: hard-limit
- hardlims: symmetric hard-limit
- satlins: satlins
- tansig: tan-sigmoid
- tribas: triangular basis
- poslin: positive linear
- purelin: linear