Briefing
Briefly about the calculations in the script. Determine the size of the ensemble n, the input Xtest and the target Ytest, using the basic set of X $ test. Then, in the foreach loop (only when the index is equal to the bestNN indexes), calculate the predicates of these best NNs, summarize them, divide by the number of the best NN. We translate the output from the continuous variable into the nominal variable (0.1) and calculate the metrics. As you can see, the quality indicators of classification are much higher than those of individual classifiers.