Classification pruning
Hence, the quality of the classification of the same ensemble on the same data, but with a different method of unification, can vary greatly.
Despite the fact that the hyperparameters of the individual classifiers of the ensemble were chosen intuitively and obviously are not optimal, we obtained a high and stable quality of classification, both using averaging and with a simple majority.