Summary
ensembles of neural network classifiers, composed of simple and fast ELM neural networks, show the quality of the classification, comparable to more complex models (DNN).

Optimization of hyperparameters of individual ensemble classifiers gives an improvement in the quality of classification to Acc = 0.77 (95% CI = 0.73 - 0.81).
The quality of the ensemble classification with averaging and voting by the majority is approximately the same.
The preservation of the quality of the ensemble classification after the training is preserved to a depth of more than half the size of the training set. In our case, the quality is maintained up to 750 bars, which is much more than the same figure obtained for DNN (250 bars).
The quality of the ensemble classification can be significantly improved by optimizing the threshold for converting the continuous predictive variable to nominal (calibration, optimal CutOff, genetic search).
The quality of the ensemble classification can be improved and using the stacking model as a combiner. It can be a neural network or an ensemble of neural networks. In the next part of the article, we will test these two variants of stacking. To build a neural network, we will test new possibilities provided by a group of TensorFlow libraries.