Bayesian Optimization
The function returns a wide range of metrics, of which now we only need the value of F1.

We will use a neural network with two hidden layers. We will conduct DNN training in two stages, with pre-training. The options are:

Pre-training:
train only SRBM;
We train SRBM + the top layer of the neural network.
Subtle training:
use the backpropagation training method;
use the training method rpropagation.
Each of these four training options has a different set of hyperparameters for optimization.